reclusters more than 2 clusters

+++ recluster.py
@@ -13,8 +13,11 @@ with open('homo_out.txt') as handle:
 for line in output:
    cluster1 = './USEARCH_results/' + line[0] 
    for name in line[1:]:
-      cluster2 = './USEARCH_results/' + name
-      cmd = 'cat ' + cluster2 + '>> ' + cluster1
-      os.system(cmd)
-      os.system('rm ' + cluster2)
+      if name == ' ':
+         continue
+      else:
+         cluster2 = './USEARCH_results/' + name
+         cmd = 'cat ' + cluster2 + '>> ' + cluster1
+         os.system(cmd)
+         os.system('rm ' + cluster2)
